requirements.txt
django==3.2.5
phonenumbers==8.12.29
requests==2.26.0
djangorestframework==3.12.4
django_filter==2.4.0
djoser==2.1.0
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0